refactor: :recycle: 폴더 구조 수정 추가
